Load SQL datastore directly
This changes the server catalog code to load and configure the SQL
datastore directly instead of going through the catalog machinery. This
removes the gRPC layer between the core and the datastore code.

Bypassing gRPC gets rid of the 4MB response limit that currently exists,
allowing us to simplify some of the interactions with the datastore.
Namely it allows us to remove the paging support added to
ListNodeSelectors (which is currently broken).

Most of the change consists of adding support to the common catalog
"Fill()" code to support a struct tag that causes the catalog code to
ignore a struct field during "filling".

Signed-off-by: Andrew Harding <aharding@vmware.com>