[WIP] Custom handlers and live and ready checks for server
This is an initial work towards making a custom HTTP handler to manage
live and ready checks for the server.

For this, two interfaces were made in the health module which can be
used by different subsystems to check if it is live or ready based on
some specific condition. The subsystems can then implement these
interfaces and then can return the state using the Status() method.

For the handler, we can check the response of these interfaces and
determine which information/state are to be returned as an output. Two
helper method to parse and format the output are to be used.

Problems:
* How to initialise the `NewChecker` function to return the interfaces
in the `Checker` struct?
* Is there some bad implementation or a better way of doing these?